Split 'on death' behavior from 'spawn fireworks' action
